+++ amd/src/incoursesettings.js
@@ -1,4 +1,3 @@
-<?php
 // This file is part of Moodle - http://moodle.org/
 //
 // Moodle is free software: you can redistribute it and/or modify
@@ -15,18 +14,30 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Campus - Version file
+ * Theme Boost Campus - JS code for displaying course setting within the course.
  *
  * @package    theme_boost_campus
  * @copyright  2017 Kathrin Osswald, Ulm University <kathrin.osswald@uni-ulm.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
+define(['jquery'], function($) {
+    "use strict";
 
-$plugin->component = 'theme_boost_campus';
-$plugin->version = 20170717801;
-$plugin->release = 'v3.2-r1';
-$plugin->requires = 2016070700;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->dependencies = array('theme_boost' => 2016102100);
+    function initInCourseSettings() {
+        $('#page-header .context-header-settings-menu').on('click', function (e) {
+            e.stopPropagation();
+            if ($('#boost-campus-course-settings').is(":visible")) {
+                $('#boost-campus-course-settings').hide(400);
+            } else {
+                $('#boost-campus-course-settings').show(400);
+            }
+        });
+    }
+
+    return {
+        init: function() {
+            initInCourseSettings();
+        }
+    };
+});